wifi: ieee80211: split EHT definitions out
[ Upstream commit 86bc0c662322b4749cd666678d2fdce7015bcae3 ]

The ieee80211.h file has gotten very long, continue splitting
it by putting EHT definitions into a separate file.

Link: https://patch.msgid.link/20251105153843.bf77fe169140.I691267e0edd914c604a5bfd447d33be00044c9b4@changeid
Stable-dep-of: cb0caadb64ca ("wifi: ieee80211: fix definition of EHT-MCS 15 in MRU")
(cherry picked from commit dd827cff429d5d20cf6a6310b6438d0d13d33028)
